Final fix for UdpChecker byte literals

+++ V2rayNG/app/src/main/java/com/samal/v2ray/handler/UdpChecker.kt
@@ -30,9 +30,15 @@ object UdpChecker {
                         soTimeout = 3000
                     }
                     val payload = if (port == 19302) {
-                        byteArrayOf(0x00, 0x01, 0x00, 0x00, 0x21, 0x12, 0xa4, 0x42, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00)
+                        byteArrayOf(
+                            0x00.toByte(), 0x01.toByte(), 0x00.toByte(), 0x00.toByte(),
+                            0x21.toByte(), 0x12.toByte(), 0xa4.toByte(), 0x42.toByte(),
+                            0x00.toByte(), 0x00.toByte(), 0x00.toByte(), 0x00.toByte(),
+                            0x00.toByte(), 0x00.toByte(), 0x00.toByte(), 0x00.toByte(),
+                            0x00.toByte(), 0x00.toByte(), 0x00.toByte(), 0x00.toByte()
+                        )
                     } else {
-                        ByteArray(48).apply { this[0] = 0x1b }
+                        ByteArray(48).apply { this[0] = 0x1b.toByte() }
                     }
 
                     val packet = DatagramPacket(payload, payload.size, addr, port)
